backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
anthropic>=0.18.0
ultralytics>=8.1.0
opencv-python>=4.9.0
python-dotenv>=1.0.1
websockets>=12.0
httpx>=0.27.0
elevenlabs
